requirements.txt
fastapi
uvicorn
python-dotenv
requests
spacy
transformers
sqlalchemy
psycopg2-binary
twilio
mysql-connector-python
faker
openai